Fix #205: Modify to use a bitmap

+++ awkernel_async_lib/src/scheduler/prioritized_fifo.rs
@@ -3,7 +3,7 @@
 use super::{Scheduler, SchedulerType, Task};
 use crate::task::State;
 use alloc::sync::Arc;
-use awkernel_lib::priority_queue::PriorityQueue;
+use awkernel_async_lib_verified::priority_queue::PriorityQueue;
 use awkernel_lib::sync::mutex::{MCSNode, Mutex};
 
 pub struct PrioritizedFIFOScheduler {
@@ -39,8 +39,9 @@ impl Scheduler for PrioritizedFIFOScheduler {
                 return;
             };
 
+            // FIXME
             data.queue.push(
-                priority as usize,
+                priority as u32,
                 PrioritizedFIFOTask {
                     task: task.clone(),
                     _priority: priority,
@@ -55,7 +56,7 @@ impl Scheduler for PrioritizedFIFOScheduler {
             };
 
             prioritized_fifo_data.queue.push(
-                priority as usize,
+                priority as u32,
                 PrioritizedFIFOTask {
                     task: task.clone(),
                     _priority: priority,
