+++ awkernel_async_lib/src/scheduler/priority_based_rr.rs
@@ -3,7 +3,7 @@
 use super::{Scheduler, SchedulerType, Task};
 use crate::task::{get_last_executed_by_task_id, set_need_preemption, State};
 use alloc::sync::Arc;
-use awkernel_lib::priority_queue::PriorityQueue;
+use awkernel_async_lib_verified::priority_queue::PriorityQueue;
 use awkernel_lib::sync::mutex::{MCSNode, Mutex};
 
 pub struct PriorityBasedRRScheduler {
@@ -46,7 +46,7 @@ impl Scheduler for PriorityBasedRRScheduler {
         let mut node = MCSNode::new();
         let mut guard = self.data.lock(&mut node);
         let data = guard.get_or_insert_with(PriorityBasedRRData::new);
-        data.queue.push(priority as usize, new_task);
+        data.queue.push(priority as u32, new_task);
     }
 
     fn get_next(&self) -> Option<Arc<Task>> {
